F13 completed - added modal view in reports page
