+++ src/Components/Modal/ModalReports/ModalReportBackdrop/ModalReportBackdrop.module.css
@@ -1,0 +1,9 @@
+.backdrop {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100vh;
+  z-index: 10;
+  background: rgba(0, 0, 0, 0.75);
+}
