+++ src/Components/Modal/ModalReports/ModalReportOverlay/ModalReportOverlay.module.css
@@ -1,0 +1,13 @@
+.modal {
+  background-color: #fff;
+  position: fixed;
+  top: 25vh;
+  left: 25%;
+  z-index: 100;
+  overflow: hidden;
+}
+
+.modal i {
+  font-size: 2.5rem;
+  cursor: pointer;
+}
