+++ src/Components/Modal/ModalReports/ModalReports.module.css
@@ -1,23 +1,1 @@
-.backdrop {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100vh;
-  z-index: 10;
-  background: rgba(0, 0, 0, 0.75);
-}
 
-.modal {
-  background-color: #fff;
-  position: fixed;
-  top: 25vh;
-  left: 25%;
-  z-index: 100;
-  overflow: hidden;
-}
-
-.modal i {
-  font-size: 2.5rem;
-  cursor: pointer;
-}
